Add email verifying on signup and on change (#19)
User's email address is now verified when signing up and when the user changes it.

If new user account is not verified it is deleted after verification token expires.

If email change is not verified before token expires user's email is not changed. User's old email address continues to be used.