Add user account deletion functionality (#23)
Implemented the ability for users to delete their accounts. Only standard users have the ability to delete their own accounts. If a user is an organization admin or super admin, only a super admin can delete that account with the exception that the only super admin account cannot be deleted.